@@ -26,29 +26,17 @@ final class RegexUtils
         $regex = preg_replace('/(?<!\\\)\*/', '{1,1}', $regex);
         $regex = preg_replace('/(?<!\\\)\+/', '{1,1}', $regex);
         // [12]{1,2} becomes [12]
-        $regex = preg_replace_callback('/(\[[^\]]+\])\{(\d+),(\d+)\}/', static function ($matches) {
-            return str_repeat($matches[1], (int) $matches[2]);
-        }, $regex);
+        $regex = preg_replace_callback('#(\[[^\]]+\])\{(\d+),(\d+)\}#', static fn ($matches): string => str_repeat($matches[1], (int) $matches[2]), $regex);
         // (12|34){1,2} becomes (12|34)
-        $regex = preg_replace_callback('/(\([^\)]+\))\{(\d+),(\d+)\}/', static function ($matches) {
-            return str_repeat($matches[1], (int) $matches[2]);
-        }, $regex ?? '');
+        $regex = preg_replace_callback('#(\([^\)]+\))\{(\d+),(\d+)\}#', static fn ($matches): string => str_repeat($matches[1], (int) $matches[2]), $regex ?? '');
         // A{1,2} becomes A or \d{3} becomes \d\d\d
-        $regex = preg_replace_callback('/(\\\?.)\{(\d+),(\d+)\}/', static function ($matches) {
-            return str_repeat($matches[1], (int) $matches[2]);
-        }, $regex ?? '');
+        $regex = preg_replace_callback('#(\\\?.)\{(\d+),(\d+)\}#', static fn ($matches): string => str_repeat($matches[1], (int) $matches[2]), $regex ?? '');
         // (this|that) becomes 'this'
-        $regex = preg_replace_callback('/\((.*?)\)/', static function ($matches) {
-            return explode('|', str_replace(['(', ')'], '', $matches[1]))[0];
-        }, $regex ?? '');
+        $regex = preg_replace_callback('#\((.*?)\)#', static fn ($matches): string => explode('|', str_replace(['(', ')'], '', $matches[1]))[0], $regex ?? '');
         // [A-F] become [A] or [0-9] becomes [0]
-        $regex = preg_replace_callback('/\[([^\]]+)\]/', static function ($matches) {
-            return '[' . preg_replace_callback('/(\w|\d)\-(\w|\d)/', static function ($range) {
-                return $range[1];
-            }, $matches[1]) . ']';
-        }, $regex ?? '');
+        $regex = preg_replace_callback('#\[([^\]]+)\]#', static fn ($matches): string => '[' . preg_replace_callback('#(\w|\d)\-(\w|\d)#', static fn ($range): string => $range[1], $matches[1]) . ']', $regex ?? '');
         // All [ABC] become A
-        $regex = preg_replace_callback('/\[([^\]]+)\]/', static function ($matches) {
+        $regex = preg_replace_callback('#\[([^\]]+)\]#', static function ($matches): string {
             // remove backslashes (that are not followed by another backslash) because they are escape characters
             $match        = preg_replace('/\\\(?!\\\)/', '', $matches[1]);
             $firstElement = str_split($match)[0];
@@ -64,8 +52,7 @@ final class RegexUtils
         // remove remaining single backslashes
         $regex = str_replace('\\\\', '[:escaped_backslash:]', $regex);
         $regex = str_replace('\\', '', $regex);
-        $regex = str_replace('[:escaped_backslash:]', '\\', $regex);
 
-        return $regex;
+        return str_replace('[:escaped_backslash:]', '\\', $regex);
     }
 }
