@@ -6,25 +6,17 @@ namespace Vural\OpenAPIFaker\Exception;
 
 use Exception;
 
-use function Safe\sprintf;
+use function sprintf;
 
-class NoExample extends Exception
+final class NoExample extends Exception
 {
-    protected string $example;
-
     public static function forRequest(string $example): self
     {
-        $e          = new self(sprintf('OpenAPI spec does not have a example "%s" request', $example));
-        $e->example = $example;
-
-        return $e;
+        return new self(sprintf('OpenAPI spec does not have an example "%s" request', $example));
     }
 
     public static function forResponse(string $example): self
     {
-        $e          = new self(sprintf('OpenAPI spec does not have a example "%s" response', $example));
-        $e->example = $example;
-
-        return $e;
+        return new self(sprintf('OpenAPI spec does not have an example "%s" response', $example));
     }
 }
