@@ -7,17 +7,23 @@ namespace Vural\OpenAPIFaker;
 use InvalidArgumentException;
 
 use function in_array;
-use function Safe\sprintf;
+use function sprintf;
 
 final class Options
 {
-    public const STRATEGY_STATIC  = 'static';
+    public const STRATEGY_STATIC = 'static';
+
     public const STRATEGY_DYNAMIC = 'dynamic';
 
-    private int|null $minItems        = null;
-    private int|null $maxItems        = null;
+    private int|null $minItems = null;
+
+    private int|null $maxItems = null;
+
     private bool $alwaysFakeOptionals = false;
-    private string $strategy          = self::STRATEGY_DYNAMIC;
+
+    private string $strategy = self::STRATEGY_DYNAMIC;
+
+    private const ALLOWED = [self::STRATEGY_STATIC, self::STRATEGY_DYNAMIC];
 
     public function setMinItems(int $minItems): Options
     {
@@ -43,9 +49,7 @@ final class Options
     /** @throws InvalidArgumentException */
     public function setStrategy(string $strategy): Options
     {
-        $allowed = [self::STRATEGY_STATIC, self::STRATEGY_DYNAMIC];
-
-        if (! in_array($strategy, $allowed, true)) {
+        if (! in_array($strategy, self::ALLOWED, true)) {
             throw new InvalidArgumentException(sprintf('Unknown generation strategy: %s', $strategy));
         }
 
