@@ -18,9 +18,9 @@ use Vural\OpenAPIFaker\Utils\StringUtils;
 use function base64_encode;
 use function max;
 use function reset;
-use function Safe\substr;
 use function str_repeat;
 use function strlen;
+use function substr;
 
 use const DATE_RFC3339;
 
@@ -71,43 +71,20 @@ final class StringFaker
         $minLength = $schema->minLength ?? 0;
         $maxLength = $schema->maxLength ?? max(10, $minLength + 1);
 
-        switch ($schema->format) {
-            case 'date':
-                return DateTime::date('Y-m-d', '2199-01-01');
-
-            case 'date-time':
-                return DateTime::dateTime('2199-01-01 00:00:00')->format(DATE_RFC3339);
-
-            case 'email':
-                return (new Internet(Factory::create()))->safeEmail();
-
-            case 'uuid':
-                return Uuid::uuid();
-
-            case 'uri':
-                return (new Internet(Factory::create()))->url();
-
-            case 'hostname':
-                return (new Internet(Factory::create()))->domainName();
-
-            case 'ipv4':
-                return (new Internet(Factory::create()))->ipv4();
-
-            case 'ipv6':
-                return (new Internet(Factory::create()))->ipv6();
-
-            case 'byte':
-                return base64_encode(Lorem::word());
-
-            case 'binary':
-                return StringUtils::convertToBinary(Lorem::word());
-
-            case 'password':
-                return Base::asciify(str_repeat('*', $maxLength));
-
-            default:
-                return Lorem::word();
-        }
+        return match ($schema->format) {
+            'date' => DateTime::date('Y-m-d', '2199-01-01'),
+            'date-time' => DateTime::dateTime('2199-01-01 00:00:00')->format(DATE_RFC3339),
+            'email' => (new Internet(Factory::create()))->safeEmail(),
+            'uuid' => Uuid::uuid(),
+            'uri' => (new Internet(Factory::create()))->url(),
+            'hostname' => (new Internet(Factory::create()))->domainName(),
+            'ipv4' => (new Internet(Factory::create()))->ipv4(),
+            'ipv6' => (new Internet(Factory::create()))->ipv6(),
+            'byte' => base64_encode(Lorem::word()),
+            'binary' => StringUtils::convertToBinary(Lorem::word()),
+            'password' => Base::asciify(str_repeat('*', $maxLength)),
+            default => Lorem::word(),
+        };
     }
 
     private static function generateStatic(Schema $schema): string|null
@@ -149,42 +126,19 @@ final class StringFaker
         $minLength = $schema->minLength;
         $maxLength = $schema->maxLength;
 
-        switch ($schema->format) {
-            case 'date':
-                return '2019-08-24';
-
-            case 'date-time':
-                return (new \Safe\DateTime('2019-08-24T14:15:22'))->format(DATE_RFC3339);
-
-            case 'email':
-                return 'user@example.com';
-
-            case 'uuid':
-                return '095be615-a8ad-4c33-8e9c-c7612fbf6c9f';
-
-            case 'uri':
-                return 'http://example.com';
-
-            case 'hostname':
-                return 'example.com';
-
-            case 'ipv4':
-                return '192.168.0.1';
-
-            case 'ipv6':
-                return '2001:0db8:85a3:0000:0000:8a2e:0370:7334';
-
-            case 'byte':
-                return base64_encode('string');
-
-            case 'binary':
-                return StringUtils::convertToBinary('string');
-
-            case 'password':
-                return StringUtils::ensureLength('pa$$wordqwerty!@#$%^123456', $minLength, $maxLength);
-
-            default:
-                return StringUtils::ensureLength('string', $minLength, $maxLength);
-        }
+        return match ($schema->format) {
+            'date' => '2019-08-24',
+            'date-time' => (new \Safe\DateTime('2019-08-24T14:15:22'))->format(DATE_RFC3339),
+            'email' => 'user@example.com',
+            'uuid' => '095be615-a8ad-4c33-8e9c-c7612fbf6c9f',
+            'uri' => 'http://example.com',
+            'hostname' => 'example.com',
+            'ipv4' => '192.168.0.1',
+            'ipv6' => '2001:0db8:85a3:0000:0000:8a2e:0370:7334',
+            'byte' => base64_encode('string'),
+            'binary' => StringUtils::convertToBinary('string'),
+            'password' => StringUtils::ensureLength('pa$$wordqwerty!@#$%^123456', $minLength, $maxLength),
+            default => StringUtils::ensureLength('string', $minLength, $maxLength),
+        };
     }
 }
