@@ -11,7 +11,7 @@ use Vural\OpenAPIFaker\Tests\SchemaFactory;
 use Vural\OpenAPIFaker\Tests\Unit\UnitTestCase;
 
 use function filter_var;
-use function Safe\sprintf;
+use function sprintf;
 use function strlen;
 
 use const FILTER_VALIDATE_URL;
